requirements.txt
alembic==1.0.1
Flask==1.0.2
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
lxml==4.2.5
requests==2.20.0
SQLAlchemy-Utils==0.33.11
